+++ src/components/dialog/dialog.js
@@ -62,46 +62,175 @@ class wux {
 			 * 默认参数
 			 */
 			defaults: {
-				showCancel: !0, 
 				title: '', 
 				content: '', 
-				confirmText: '确定', 
-				cancelText: '取消', 
-				confirm: function() {}, 
-				cancel: function() {}, 
+				buttons: [], 
+			},
+			/**
+			 * 默认数据
+			 */
+			data() {
+				return {
+					onCancel: function() {}, 
+		        	cancelText: '取消', 
+		        	cancelType: 'weui-dialog__btn_default', 
+		        	onConfirm: function() {}, 
+		        	confirmText: '确定', 
+		        	confirmType: 'weui-dialog__btn_primary', 
+		        }
 			},
 			/**
 			 * 显示dialog组件
 			 * @param {Object} opts 参数对象
-			 * @param {Boolean} opts.showCancel 是否显示取消按钮
 			 * @param {String} opts.title 提示标题
 			 * @param {String} opts.content 提示文本
-			 * @param {String} opts.confirmText 确定按钮的文字
-			 * @param {String} opts.cancelText 取消按钮的文字
-			 * @param {Function} opts.confirm 点击确定按钮的回调函数
-			 * @param {Function} opts.cancel 点击按钮按钮的回调函数
+			 * @param {Array} opts.buttons 按钮
+			 * @param {String} opts.buttons.text 按钮的文字
+			 * @param {String} opts.buttons.type 按钮的类型
+			 * @param {Function} opts.buttons.onTap 按钮的点击事件
 			 */
 			open(opts = {}) {
 				const options = extend(clone(this.defaults), opts)
-				const hideDialog = (cb) => {
+				const self = {}
+
+				// 隐藏
+				self.hide = (cb) => {
+					if (self.removed) return !1
+					self.removed = !0
 					that.setHidden('dialog')
-					typeof cb === 'function' && cb()
+					setTimeout(() => typeof cb === 'function' && cb(), 300)
+				}
+
+				// 显示
+				self.show = () => {
+					if (self.removed) return !1
+					that.setVisible('dialog')
 				}
 
 				// 渲染组件
 				$scope.setData({
 					[`$wux.dialog`]: options, 
-					[`$wux.dialog.dialogConfirm`]: `dialogConfirm`, 
-					[`$wux.dialog.dialogCancel`]: `dialogCancel`, 
+					[`$wux.dialog.buttonTapped`]: `dialogButtonTapped`, 
 				})
 
 				// 绑定tap事件
-				$scope.dialogConfirm = () => hideDialog(options.confirm)
-				$scope.dialogCancel = () => hideDialog(options.cancel)
+				$scope[`dialogButtonTapped`] = (e) => {
+					const index = e.currentTarget.dataset.index
+					const button = options.buttons[index]
+					self.hide(() => typeof button.onTap === 'function' && button.onTap(e))
+				}
 
-				that.setVisible('dialog')
+				// 绑定input事件
+				$scope[`dialogBindinput`] = (e) => {
+					$scope.setData({
+						[`$wux.dialog.prompt.response`]: e.detail.value, 
+					})
+				}
 
-				return $scope.dialogCancel
+				self.show()
+
+				return self.hide
+			},
+			/**
+			 * 显示dialog组件
+			 * @param {Object} opts 参数对象
+			 * @param {String} opts.title 提示标题
+			 * @param {String} opts.content 提示文本
+			 * @param {String} opts.confirmText 确定按钮的文字，默认为"确定"
+			 * @param {String} opts.confirmType 确定按钮的类型
+			 * @param {Function} opts.onConfirm 确定按钮的点击事件
+			 */
+			alert(opts = {}) {
+				return this.open(extend({
+					buttons: [
+						{
+							text: opts.confirmText || this.data().confirmText, 
+							type: opts.confirmType || this.data().confirmType, 
+							onTap(e) {
+								typeof opts.onConfirm === 'function' && opts.onConfirm(e)
+							},
+						},
+					],
+				}, opts))
+			},
+			/**
+			 * 显示dialog组件
+			 * @param {Object} opts 参数对象
+			 * @param {String} opts.title 提示标题
+			 * @param {String} opts.content 提示文本
+			 * @param {String} opts.confirmText 确定按钮的文字，默认为"确定"
+			 * @param {String} opts.confirmType 确定按钮的类型
+			 * @param {Function} opts.onConfirm 确定按钮的点击事件
+			 * @param {String} opts.cancelText 取消按钮的文字，默认为"取消"
+			 * @param {String} opts.cancelType 取消按钮的类型
+			 * @param {Function} opts.onCancel 取消按钮的点击事件
+			 */
+			confirm(opts = {}) {
+				return this.open(extend({
+					buttons: [
+						{
+							text: opts.cancelText || this.data().cancelText, 
+							type: opts.cancelType || this.data().cancelType, 
+							onTap(e) { 
+								typeof opts.onCancel === 'function' && opts.onCancel(e)
+							},
+						},
+						{
+							text: opts.confirmText || this.data().confirmText, 
+							type: opts.confirmType || this.data().confirmType, 
+							onTap(e) { 
+								typeof opts.onConfirm === 'function' && opts.onConfirm(e)
+							},
+						},
+					],
+				}, opts))
+			},
+			/**
+			 * 显示dialog组件
+			 * @param {Object} opts 参数对象
+			 * @param {String} opts.title 提示标题
+			 * @param {String} opts.content 提示文本
+			 * @param {String} opts.fieldtype input 的类型，有效值：text, number, idcard, digit
+			 * @param {Boolean} opts.password 是否是密码类型
+			 * @param {String} opts.defaultText 默认值
+			 * @param {String} opts.placeholder 输入框为空时占位符
+			 * @param {Number} opts.maxlength 最大输入长度，设置为 -1 的时候不限制最大长度
+			 * @param {String} opts.confirmText 确定按钮的文字，默认为"确定"
+			 * @param {String} opts.confirmType 确定按钮的类型
+			 * @param {Function} opts.onConfirm 确定按钮的点击事件
+			 * @param {String} opts.cancelText 取消按钮的文字，默认为"取消"
+			 * @param {String} opts.cancelType 取消按钮的类型
+			 * @param {Function} opts.onCancel 取消按钮的点击事件
+			 */
+			prompt(opts = {}) {
+				const prompt = {
+					fieldtype: opts.fieldtype ? opts.fieldtype : 'text', 
+					password: !!opts.password, 
+					response: opts.defaultText ? opts.defaultText : '', 
+					placeholder: opts.placeholder ? opts.placeholder : '', 
+					maxlength: opts.maxlength ? parseInt(opts.maxlength) : '', 
+					bindinput: `dialogBindinput`
+				}
+
+				return this.open(extend({
+					prompt: prompt, 
+					buttons: [
+						{
+							text: opts.cancelText || this.data().cancelText, 
+							type: opts.cancelType || this.data().cancelType, 
+							onTap(e) { 
+								typeof opts.onCancel === 'function' && opts.onCancel(e)
+							},
+						},
+						{
+							text: opts.confirmText || this.data().confirmText, 
+							type: opts.confirmType || this.data().confirmType, 
+							onTap(e) { 
+								typeof opts.onConfirm === 'function' && opts.onConfirm(e)
+							},
+						},
+					],
+				}, opts))
 			},
 		}
 	}