+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### v 1.0.2 (2017-02-16)
+
+- 【优化】 dialog 组件，增加 open、confirm、alert、propmt 快速调用方法
+- 【优化】 gallery 组件，更新动画方式为 slideRight
+- 【修复】 actionsheet 组件，延时 hide 调用之后二次调用会触发组件隐藏
+
 #### v 1.0.1 (2017-02-15)
 
 - 【增强】 增加 barcode 组件